refactor: extract bp_health to routes/health.py
Phase 5.5 of the incremental modularisation. Moves the 11 endpoints
registered on bp_health into routes/health.py: /api/reliability,
/api/heatmap, /api/system-health, /api/health, /api/diagnostics,
/api/service-status, /api/heartbeat-status, /api/heartbeat-ping,
/api/rate-limits, /api/health-stream (SSE), /api/sandbox-status.

Pure mechanical move — zero behaviour change. Interleaved helpers
(_detect_channel_status, _DEFAULT_RATE_LIMITS, _infer_provider) stay
in dashboard.py and are reached from the route module via late
"import dashboard as _d".

dashboard.py shrinks by ~860 lines (31260 -> 30397); routes/health.py
is 918 lines.
